requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
django-js-asset==2.0.0
django-razorpay==1.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
filelock==3.10.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
payu-python==0.1.5
payu-sdk==1.2.0
Pillow==9.4.0
pipenv==2023.3.20
platformdirs==3.2.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python-http-client==3.3.7
pytz==2022.7.1
razorpay==1.3.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.21.0
virtualenv-clone==0.5.7
